Update Lab7.py
This is a weather application. requests plug in must be installed to work. Basically run and it will ask you for a city. Enter a city name is it will information about the city. This is explained in comments through the code. The multiple uses are that it displays multiple forms of information. Weather, Coordinates, Wind, Visibility, Timezones etc.